requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
Pillow
